Exercicio_pratico_HIbernate_20230802

+++ src/main/java/br/unb/dao/AlunoDAO.java
@@ -24,20 +24,6 @@ public class AlunoDAO {
 
 	public void salvar(Aluno aluno) throws SQLException {
 		
-//		try (Connection connection = DriverManager.getConnection(URL, USER, PASSWORD)) {
-//			String sql = "INSERT INTO aluno (nome, matricula) VALUES (?, ?)";
-//			try (PreparedStatement statement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
-//				statement.setString(1, aluno.getNome());
-//				statement.setString(2, aluno.getMatricula());
-//				statement.executeUpdate();
-//
-//				ResultSet generatedKeys = statement.getGeneratedKeys();
-//				if (generatedKeys.next()) {
-//					aluno.setId(generatedKeys.getInt(1));
-//				}
-//			}
-//		}
-		
 		SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
 		Session session = sessionFactory.openSession();
 
@@ -50,15 +36,6 @@ public class AlunoDAO {
 
 	public void atualizar(Aluno aluno) throws SQLException {
 		System.out.println("AlunoDAO::atualizar::aluno->"+aluno);
-//		try (Connection connection = DriverManager.getConnection(URL, USER, PASSWORD)) {
-//			String sql = "UPDATE aluno SET nome = ?, matricula = ? WHERE id = ?";
-//			try (PreparedStatement statement = connection.prepareStatement(sql)) {
-//				statement.setString(1, aluno.getNome());
-//				statement.setString(2, aluno.getMatricula());
-//				statement.setInt(3, aluno.getId());
-//				statement.executeUpdate();
-//			}
-//		}
 		
 		Session session = HibernateUtil.getSessionFactory().openSession();
 		Transaction tx = session.beginTransaction();
@@ -69,14 +46,7 @@ public class AlunoDAO {
 		session.close();
 	}
 
-	public void excluir(int id) throws SQLException {
-//		try (Connection connection = DriverManager.getConnection(URL, USER, PASSWORD)) {
-//			String sql = "DELETE FROM aluno WHERE id = ?";
-//			try (PreparedStatement statement = connection.prepareStatement(sql)) {
-//				statement.setInt(1, id);
-//				statement.executeUpdate();
-//			}
-//		}
+	public void excluir(Long id) throws SQLException {
 		System.out.println("AlunoDAO::excluir::aluno->"+id);
 		Session session = HibernateUtil.getSessionFactory().openSession();
 		Transaction tx = session.beginTransaction();
@@ -88,22 +58,7 @@ public class AlunoDAO {
 		session.close();
 	}
 
-	public Aluno buscarPorId(int id) throws SQLException {
-//		try (Connection connection = DriverManager.getConnection(URL, USER, PASSWORD)) {
-//			String sql = "SELECT * FROM aluno WHERE id = ?";
-//			try (PreparedStatement statement = connection.prepareStatement(sql)) {
-//				statement.setInt(1, id);
-//				try (ResultSet resultSet = statement.executeQuery()) {
-//					if (resultSet.next()) {
-//						String nome = resultSet.getString("nome");
-//						String matricula = resultSet.getString("matricula");
-//						Aluno aluno = new Aluno(nome, matricula);
-//						aluno.setId(id);
-//						return aluno;
-//					}
-//				}
-//			}
-//		}
+	public Aluno buscarPorId(Long id) throws SQLException {
 		
 		Session session = HibernateUtil.getSessionFactory().openSession();
 
@@ -115,22 +70,6 @@ public class AlunoDAO {
 
 	public List<Aluno> listarTodos() throws SQLException, ClassNotFoundException {
 		System.out.println("AlunoDAO::listarTodos");
-//		List<Aluno> alunos = new ArrayList<>();
-//		try (Connection connection = DriverManager.getConnection(URL, USER, PASSWORD)) {
-//			String sql = "SELECT * FROM aluno";
-//			try (PreparedStatement statement = connection.prepareStatement(sql)) {
-//				try (ResultSet resultSet = statement.executeQuery()) {
-//					while (resultSet.next()) {
-//						int id = resultSet.getInt("id");
-//						String nome = resultSet.getString("nome");
-//						String matricula = resultSet.getString("matricula");
-//						Aluno aluno = new Aluno(nome, matricula);
-//						aluno.setId(id);
-//						alunos.add(aluno);
-//					}
-//				}
-//			}
-//		}
 		
 		Session session = HibernateUtil.getSessionFactory().openSession();
 		Query query = session.createQuery("FROM Aluno");
